requirements.txt
pillow
pytesseract
langdetect
googletrans==4.0.0-rc1
opencv-python